Rename the test packages to test only the public APIs
By keeping the test classes in the same package I can reference a field
with protected or default visibility that is not available publicly. In
this case it is useful to test only the public APIs to ensure all
relevant classes are available, nothing else.
