+++ src/main/java/org/jscookie/CookieSerializationException.java
@@ -1,6 +1,6 @@
 package org.jscookie;
 
-class CookieSerializationException extends Exception {
+public class CookieSerializationException extends Exception {
 	private static final long serialVersionUID = 1;
 	CookieSerializationException( Throwable cause ) {
 		super( cause );
