+++ src/test/java/org/jscookie/test/unit/CookiesJSONWriteTest.java
@@ -1,10 +1,13 @@
-package org.jscookie;
+package org.jscookie.test.unit;
 
 import java.util.Arrays;
 
 import javax.servlet.http.Cookie;
 
-import org.jscookie.testutils.BaseTest;
+import org.jscookie.CookieSerializationException;
+import org.jscookie.CookieValue;
+import org.jscookie.Cookies;
+import org.jscookie.test.unit.utils.BaseTest;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
